+++ src/main/java/com/portalSite/acquisition/dto/kafkaDto/KeywordScore.java
@@ -1,4 +1,4 @@
 package com.portalSite.acquisition.dto.kafkaDto;
 
-public record KeywordScore(String keyword, long score) {
+public record KeywordScore(String keyword, double score) {
 }
